encapsulate react-dnd Drop logic into its own file, vastly simplify QuadTile component rendering code
